Extract PEM key handling into a utility class
