+++ src/main/java/com/spotify/github/v3/clients/JwtTokenIssuer.java
@@ -30,18 +30,13 @@ import java.security.PrivateKey;
 import java.security.spec.InvalidKeySpecException;
 import java.security.spec.KeySpec;
 import java.security.spec.PKCS8EncodedKeySpec;
-import java.util.Base64;
 import java.util.Date;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
 import org.apache.commons.io.FileUtils;
 
 /** The helper Jwt token issuer. */
 public class JwtTokenIssuer {
 
   private static final SignatureAlgorithm SIGNATURE_ALGORITHM = SignatureAlgorithm.RS256;
-  private static final Pattern PKCS1_PEM_KEY_PATTERN =
-      Pattern.compile("(?m)(?s)^---*BEGIN RSA PRIVATE KEY.*---*$(.*)^---*END.*---*$.*");
   private static final long TOKEN_TTL = 600000;
 
   private final PrivateKey signingKey;
@@ -74,18 +69,11 @@ public class JwtTokenIssuer {
   public static JwtTokenIssuer fromPrivateKey(final byte[] privateKey)
       throws NoSuchAlgorithmException, InvalidKeySpecException {
 
-    final KeySpec spec;
-    final Matcher isPEM = PKCS1_PEM_KEY_PATTERN.matcher(new String(privateKey));
-    if (isPEM.matches()) {
-      byte[] pkcs1Key = Base64.getMimeDecoder().decode(isPEM.group(1));
-      byte[] pkcs8Key = toPkcs8(pkcs1Key);
-      spec = new PKCS8EncodedKeySpec(pkcs8Key);
-    } else {
-      spec = new PKCS8EncodedKeySpec(privateKey);
-    }
+    KeySpec keySpec = PKCS1PEMKey.loadKeySpec(privateKey)
+        .orElseGet(() -> new PKCS8EncodedKeySpec(privateKey));
 
     KeyFactory kf = KeyFactory.getInstance("RSA");
-    PrivateKey signingKey = kf.generatePrivate(spec);
+    PrivateKey signingKey = kf.generatePrivate(keySpec);
     return new JwtTokenIssuer(signingKey);
   }
 
@@ -104,33 +92,5 @@ public class JwtTokenIssuer {
         .setExpiration(new Date(System.currentTimeMillis() + TOKEN_TTL))
         .setIssuedAt(new Date())
         .compact();
-  }
-
-  /**
-   * Convert a PKCS#1 key to a PKCS#8 key.
-   *
-   * <p>The Github app key comes in PKCS#1 format, while the Java security utilities only natively
-   * understand PKCS#8. Fortunately, we can convert between the two by adding the PKCS#8 headers
-   * manually.
-   *
-   * <p>Adapted from {@link https://github.com/Mastercard/client-encryption-java/blob/master/src
-   * /main/java/com/mastercard/developer/utils/EncryptionUtils.java}
-   */
-  @SuppressWarnings("checkstyle:magicnumber")
-  //checkstyle:off
-  private static byte[] toPkcs8(final byte[] pkcs1Bytes) {
-    final int pkcs1Length = pkcs1Bytes.length;
-    final int totalLength = pkcs1Length + 22;
-    byte[] pkcs8Header = new byte[] {
-        0x30, (byte) 0x82, (byte) ((totalLength >> 8) & 0xff), (byte) (totalLength & 0xff), // Sequence + total length
-        0x2, 0x1, 0x0, // Integer (0)
-        0x30, 0xD, 0x6, 0x9, 0x2A, (byte) 0x86, 0x48, (byte) 0x86, (byte) 0xF7, 0xD, 0x1, 0x1, 0x1, 0x5, 0x0, // Sequence: 1.2.840.113549.1.1.1, NULL
-        0x4, (byte) 0x82, (byte) ((pkcs1Length >> 8) & 0xff), (byte) (pkcs1Length & 0xff) // Octet string + length
-    };
-
-    byte[] pkcs8bytes = new byte[pkcs8Header.length + pkcs1Bytes.length];
-    System.arraycopy(pkcs8Header, 0, pkcs8bytes, 0, pkcs8Header.length);
-    System.arraycopy(pkcs1Bytes, 0, pkcs8bytes, pkcs8Header.length, pkcs1Bytes.length);
-    return pkcs8bytes;
   }
 }
